+++ src/convert-3d-circuit-to-svg.tsx
@@ -1,17 +1,12 @@
-import jscad from "@jscad/modeling"
 import type { AnySoupElement } from "@tscircuit/soup"
 import { su } from "@tscircuit/soup-util"
-import { Footprinter3d } from "jscad-electronics"
-import { convertCSGToThreeGeom, createJSCADRenderer } from "jscad-fiber"
-import { executeJscadOperations, jscadPlanner } from "jscad-planner"
 import { JSDOM } from "jsdom"
 import * as THREE from "three"
-import { BufferGeometry, Float32BufferAttribute } from "three"
-import { OBJLoader } from "three/examples/jsm/loaders/OBJLoader.js"
-import { STLLoader } from "three/examples/jsm/loaders/STLLoader.js"
 import { SVGRenderer } from "three/examples/jsm/renderers/SVGRenderer.js"
 import { createBoardGeomFromSoup } from "./soup-to-3d"
 import { applyJsdomShim } from "./utils/jsdom-shim"
+import { createGeometryFromPolygons } from "./utils/create-geometry-from-polygons"
+import { renderComponent } from "./utils/render-component"
 
 interface CircuitToSvgOptions {
   width?: number
@@ -31,79 +26,6 @@ interface CircuitToSvgOptions {
       y: number
       z: number
     }
-  }
-}
-
-const { createJSCADRoot } = createJSCADRenderer(jscadPlanner as any)
-
-function createGeometryFromPolygons(polygons: any[]) {
-  const geometry = new BufferGeometry()
-  const vertices: number[] = []
-  const normals: number[] = []
-
-  try {
-    for (const polygon of polygons) {
-      // Create triangles from polygon vertices
-      for (let i = 1; i < polygon.vertices.length - 1; i++) {
-        vertices.push(
-          ...polygon.vertices[0], // First vertex
-          ...polygon.vertices[i], // Second vertex
-          ...polygon.vertices[i + 1], // Third vertex
-        )
-
-        // Add normal for each vertex of the triangle
-        const v1 = new THREE.Vector3(...polygon.vertices[0])
-        const v2 = new THREE.Vector3(...polygon.vertices[i])
-        const v3 = new THREE.Vector3(...polygon.vertices[i + 1])
-        const normal = new THREE.Vector3()
-          .crossVectors(
-            new THREE.Vector3().subVectors(v2, v1),
-            new THREE.Vector3().subVectors(v3, v1),
-          )
-          .normalize()
-
-        normals.push(
-          normal.x,
-          normal.y,
-          normal.z,
-          normal.x,
-          normal.y,
-          normal.z,
-          normal.x,
-          normal.y,
-          normal.z,
-        )
-      }
-    }
-
-    geometry.setAttribute("position", new Float32BufferAttribute(vertices, 3))
-    geometry.setAttribute("normal", new Float32BufferAttribute(normals, 3))
-    return geometry
-  } catch (error) {
-    console.error("Error creating geometry:", error)
-    throw error
-  }
-}
-
-async function loadModel(url: string): Promise<THREE.Object3D | null> {
-  try {
-    if (url.endsWith(".stl")) {
-      const loader = new STLLoader()
-      const geometry = await loader.loadAsync(url)
-      const material = new THREE.MeshStandardMaterial({
-        color: 0x888888,
-        metalness: 0.5,
-        roughness: 0.5,
-      })
-      return new THREE.Mesh(geometry, material)
-    } else if (url.endsWith(".obj")) {
-      const loader = new OBJLoader()
-      return await loader.loadAsync(url)
-    }
-    return null
-  } catch (error) {
-    console.error("Failed to load model:", error)
-    return null
   }
 }
 
@@ -138,7 +60,6 @@ export async function convert3dCircuitToSvg(
     1000,
   )
 
-  // Use camera position from options if provided
   if (options.camera?.position) {
     camera.position.set(
       options.camera.position.x,
@@ -151,7 +72,6 @@ export async function convert3dCircuitToSvg(
 
   camera.up.set(0, 1, 0)
 
-  // Use lookAt from options if provided
   if (options.camera?.lookAt) {
     camera.lookAt(
       options.camera.lookAt.x,
@@ -172,126 +92,7 @@ export async function convert3dCircuitToSvg(
   // Add components
   const components = su(circuitJson).cad_component.list()
   for (const component of components) {
-    try {
-      // Handle STL/OBJ models first
-      const url = component.model_obj_url ?? component.model_stl_url
-      if (url) {
-        const model = await loadModel(url)
-        if (model) {
-          if (component.position) {
-            model.position.set(
-              component.position.x ?? 0,
-              component.position.y ?? 0,
-              (component.position.z ?? 0) + 0.5,
-            )
-          }
-          if (component.rotation) {
-            model.rotation.set(
-              THREE.MathUtils.degToRad(component.rotation.x ?? 0),
-              THREE.MathUtils.degToRad(component.rotation.y ?? 0),
-              THREE.MathUtils.degToRad(component.rotation.z ?? 0),
-            )
-          }
-          scene.add(model)
-          continue
-        }
-      }
-
-      // Handle JSCAD models
-      if (component.model_jscad) {
-        const jscadObject = executeJscadOperations(
-          jscad as any,
-          component.model_jscad,
-        )
-        const threeGeom = convertCSGToThreeGeom(jscadObject)
-        const material = new THREE.MeshStandardMaterial({
-          color: 0x888888,
-          metalness: 0.5,
-          roughness: 0.5,
-          side: THREE.DoubleSide,
-        })
-        const mesh = new THREE.Mesh(threeGeom, material)
-
-        if (component.position) {
-          mesh.position.set(
-            component.position.x ?? 0,
-            component.position.y ?? 0,
-            (component.position.z ?? 0) + 0.5,
-          )
-        }
-        if (component.rotation) {
-          mesh.rotation.set(
-            THREE.MathUtils.degToRad(component.rotation.x ?? 0),
-            THREE.MathUtils.degToRad(component.rotation.y ?? 0),
-            THREE.MathUtils.degToRad(component.rotation.z ?? 0),
-          )
-        }
-        scene.add(mesh)
-        continue
-      }
-
-      // Handle footprints
-      if (component.footprinter_string) {
-        try {
-          const jscadOperations: any[] = []
-          const root = createJSCADRoot(jscadOperations)
-          root.render(
-            <Footprinter3d footprint={component.footprinter_string} />,
-          )
-
-          // Process each operation from the footprinter
-          for (const operation of jscadOperations) {
-            const jscadObject = executeJscadOperations(jscad as any, operation)
-            const threeGeom = convertCSGToThreeGeom(jscadObject)
-            const material = new THREE.MeshStandardMaterial({
-              color: 0x444444, // Dark gray like in CadViewer
-              metalness: 0.2,
-              roughness: 0.8,
-              side: THREE.DoubleSide,
-            })
-            const mesh = new THREE.Mesh(threeGeom, material)
-
-            if (component.position) {
-              mesh.position.set(
-                component.position.x ?? 0,
-                component.position.y ?? 0,
-                (component.position.z ?? 0) + 0.5,
-              )
-            }
-            if (component.rotation) {
-              mesh.rotation.set(
-                THREE.MathUtils.degToRad(component.rotation.x ?? 0),
-                THREE.MathUtils.degToRad(component.rotation.y ?? 0),
-                THREE.MathUtils.degToRad(component.rotation.z ?? 0),
-              )
-            }
-            scene.add(mesh)
-          }
-        } catch (error) {
-          console.error("Failed to create footprint geometry:", error)
-        }
-      }
-    } catch (error) {
-      console.error("Failed to create component:", error)
-
-      // Add fallback box for failed components (like MixedStlModel does)
-      const geometry = new THREE.BoxGeometry(0.5, 0.5, 0.5)
-      const material = new THREE.MeshStandardMaterial({
-        color: 0xff0000,
-        transparent: true,
-        opacity: 0.25,
-      })
-      const mesh = new THREE.Mesh(geometry, material)
-
-      if (component.position) {
-        mesh.position.set(
-          component.position.x ?? 0,
-          component.position.y ?? 0,
-          (component.position.z ?? 0) + 0.5,
-        )
-      }
-      scene.add(mesh)
-    }
+    await renderComponent(component, scene)
   }
 
   // Add board geometry after components
@@ -300,7 +101,6 @@ export async function convert3dCircuitToSvg(
     for (const geom of boardGeom) {
       try {
         const geometry = createGeometryFromPolygons(geom.polygons)
-
         const material = new THREE.MeshStandardMaterial({
           color: new THREE.Color(
             geom.color?.[0] ?? 0,
@@ -309,11 +109,10 @@ export async function convert3dCircuitToSvg(
           ),
           metalness: 0.1,
           roughness: 0.8,
-          opacity: 0.9, // Slightly more transparent
+          opacity: 0.9,
           transparent: true,
           side: THREE.DoubleSide,
         })
-
         const mesh = new THREE.Mesh(geometry, material)
         scene.add(mesh)
       } catch (error) {
@@ -322,7 +121,7 @@ export async function convert3dCircuitToSvg(
     }
   }
 
-  // Add grid (matching CadViewerContainer)
+  // Add grid
   const gridHelper = new THREE.GridHelper(100, 100)
   gridHelper.rotation.x = Math.PI / 2
   scene.add(gridHelper)
@@ -343,12 +142,10 @@ export async function convert3dCircuitToSvg(
   // Render and return SVG
   renderer.render(scene, camera)
 
-  const svgString = new dom.window.XMLSerializer()
+  return new dom.window.XMLSerializer()
     .serializeToString(renderer.domElement)
     .replace(
       /xmlns="[^"]*"\s+xmlns="[^"]*"/,
       'xmlns="http://www.w3.org/2000/svg"',
     )
-
-  return svgString
 }
